requirements.txt
pytest==7.0.1
selenium==4.1.0
webdriver-manager==3.5.3
Faker==13.2.0
allure-pytest==2.9.45